Clear input text area after being submitted
The text submitted stayed in the input field. The user had to delete it manually before being able to write down a new name.

I wrote one line of code inside the add function. This code resets the value after being pushed to the queue. The result is an empty input ready for a new name.