Avoid adding empty places in queue and clear input text area with Fast Track button
When the user clicked the Fast Track button without filling anything, the app created an empty ticket in line.
When the user added a new person with the Fast Track button, the text remained in the input field. The user had to delete it manually before being able to write down a new name.

I took the code from the add() function and added it in the fastTrack() function

Since both functions are using very similar code, there might be a need to refactor it.